Use only one autoConf class
