+++ daikon-spring/daikon-spring-audit-logs/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.talend.daikon.spring.audit.logs.config.AuditLogAutoConfiguration,\
-org.talend.daikon.spring.audit.logs.config.AuditLogWebAutoConfiguration
+org.talend.daikon.spring.audit.logs.config.AuditLogAutoConfiguration
